Extract runner-viewers region into sections/runner-viewers.js
Biggest single extraction of P-09: ~490 LOC. Owns _runnerViewers (previously exported from dashboard-main.js for reload-guards to consume; reload-guards.js now imports it from the new module, breaking one dashboard-main.js edge).

The mid-region code tab · "open with" popover (gitToggleOpenPopover / gitClosePopovers + their document click/keydown listeners) was textually inside runner-viewers but unrelated — moved into sections/git-actions.js where it belongs.

The IIFE that monkey-patches window._reloadInPlace / window.reloadNode is preserved verbatim inside initRunnerViewersSideEffects(). It's been dead code since the ESM migration (the module-scoped imports are separate bindings from window.X) but fixing it is out of refactor scope — the explicit _runnerRefreshRepoNode calls from runnerStart/Stop/ForceStop are what actually keeps viewers reattached today.
